requirements.txt
gradio
opencv-python
numpy
Pillow
httpx
requests
idna
